feat(places): export expanded data tags to kml in places exporter
