+++ src/plugin/places/placesmenu.js
@@ -33,7 +33,9 @@ const KMLLayerNode = goog.require('plugin.file.kml.ui.KMLLayerNode');
 const KMLNode = goog.require('plugin.file.kml.ui.KMLNode');
 const KMLTreeExportUI = goog.require('plugin.file.kml.ui.KMLTreeExportUI');
 const places = goog.require('plugin.places');
+const EventType = goog.require('plugin.places.EventType');
 const PlacesManager = goog.require('plugin.places.PlacesManager');
+const PlacesUI = goog.require('plugin.places.ui.PlacesUI');
 const QuickAddPlacesUI = goog.require('plugin.places.ui.QuickAddPlacesUI');
 const {launchSavePlaces} = goog.require('plugin.places.ui.launchSavePlaces');
 
@@ -45,27 +47,6 @@ const {FolderOptions, PlacemarkOptions} = goog.requireType('plugin.file.kml.ui')
  * @type {string}
  */
 const GROUP_LABEL = places.TITLE;
-
-/**
- * Places menu event types.
- * @enum {string}
- */
-const EventType = {
-  SAVE_TO: 'places:saveToPlaces',
-  SAVE_TO_ANNOTATION: 'places:saveToAnnotation',
-  EXPORT: 'places:export',
-
-  // create/edit
-  ADD_FOLDER: 'places:addFolder',
-  ADD_ANNOTATION: 'places:addAnnotation',
-  ADD_PLACEMARK: 'places:addPlacemark',
-  EDIT_FOLDER: 'places:editFolder',
-  EDIT_PLACEMARK: 'places:editPlacemark',
-  QUICK_ADD_PLACES: 'places:quickAdd',
-  FEATURE_LIST: 'places:featureList',
-  REMOVE_PLACE: 'places:removePlace',
-  REMOVE_ALL: 'places:removeAll'
-};
 
 /**
  * Add places items to the layer menu.
@@ -571,8 +552,7 @@ const onLayerEvent_ = function(event) {
             QuickAddPlacesUI.launch();
             break;
           case EventType.EXPORT:
-            KMLTreeExportUI.launchTreeExport(/** @type {!KMLNode} */
-                (rootNode), 'Export Places');
+            PlacesUI.launchExportUI();
             break;
           case EventType.REMOVE_ALL:
             var children = /** @type {Array<!KMLNode>} */ (node.getChildren());
@@ -910,7 +890,6 @@ const copyNode_ = function(node) {
 
 exports = {
   GROUP_LABEL,
-  EventType,
   layerSetup,
   layerDispose,
   mapSetup,
